chore: reorganize carousel components
