+++ website-frontend/src/lib/components/carousels/LandingHero.svelte
